requirements.txt
torchvision
numpy
scikit-learn
terminaltables
pandas
scipy
ruamel.yaml
